Finishing touches for big_noritake and split screen
