+++ include/devices/led_clock.h
@@ -3,6 +3,7 @@
 #include <hal/gpio_hal.h>
 #include <input/keypad.h>
 #include <input/infrared.h>
+#include <devices/ir_remotes/all_remotes.h>
 
 #define HAS_OUTPUT_AKIZUKI_K875
 #define HAS_KEYPAD
@@ -45,24 +46,4 @@ const keypad_definition_t HWCONF_KEYPAD = {
     {GPIO_NUM_34, KEY_DOWN},
     {GPIO_NUM_27, KEY_UP},
     {GPIO_NUM_35, KEY_HEADPAT}, //<- finally a clock to have hardware headpat!
-};
-
-const infrared_definition_t HWCONF_IR_BUTTONS = {
-    /**
-        Optosupply OE13KIR (https://akizukidenshi.com/goodsaffix/OE13KIR.pdf)
-    */
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0xA0, .value = 0x8F705FA, .key = KEY_UP}, // UP
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x00, .value = 0x8F700FF, .key = KEY_DOWN}, // DOWN
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x10, .value = 0x8F708F7, .key = KEY_LEFT}, // LEFT
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x80, .value = 0x8F701FE, .key = KEY_RIGHT}, // RIGHT
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x20, .value = 0x8F704FB, .key = KEY_HEADPAT}, // CENTER
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0xf8, .value = 0x8f71fe0, .key = KEY_SOFT_F1}, // A
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x78, .value = 0x8f71ee1, .key = KEY_SOFT_F2}, // B
-    { .protocol = IRPROTO_NEC, .address = 0x10, .command = 0x58, .value = 0x8f71ae5, .key = KEY_SOFT_F3}, // C
-
-    // Top-left: TYPE=3, VALUE=0x8f78d72, ADDRESS=0x10, COMMAND=0xb1
-    // Top-right: TYPE=3, VALUE=0x8f7847b, ADDRESS=0x10, COMMAND=0x21
-    // Bottom-left: TYPE=3, VALUE=0x8f78877, ADDRESS=0x10, COMMAND=0x11
-    // Bottom-right: TYPE=3, VALUE=0x8f7817e, ADDRESS=0x10, COMMAND=0x81
-    // Power: TYPE=3, VALUE=0x8f71be4, ADDRESS=0x10, COMMAND=0xd8
 };
