+++ include/views/transitions/transitions.h
@@ -15,18 +15,12 @@ typedef enum transition_type: uint8_t {
 
 
 static Transition* transition_type_to_transition(transition_type_t type) {  
-    static WipeTransition wipe_transition = WipeTransition();
-    static SlideTransition h_slide_transition = SlideTransition(false);
-    static SlideTransition rev_h_slide_transition = SlideTransition(true);
-    static VerticalSlideTransition v_slide_transition = VerticalSlideTransition(false);
-    static VerticalSlideTransition rev_v_slide_transition = VerticalSlideTransition(true);
-
     switch(type) {
-        case TRANSITION_WIPE: return &wipe_transition;
-        case TRANSITION_SLIDE_HORIZONTAL_LEFT:  return &h_slide_transition;
-        case TRANSITION_SLIDE_HORIZONTAL_RIGHT:  return &rev_h_slide_transition;
-        case TRANSITION_SLIDE_VERTICAL_UP: return &v_slide_transition;
-        case TRANSITION_SLIDE_VERTICAL_DOWN: return &rev_v_slide_transition;
+        case TRANSITION_WIPE: return new WipeTransition();
+        case TRANSITION_SLIDE_HORIZONTAL_LEFT:  return new SlideTransition(false);
+        case TRANSITION_SLIDE_HORIZONTAL_RIGHT:  return new SlideTransition(true);
+        case TRANSITION_SLIDE_VERTICAL_UP: return new VerticalSlideTransition(false);
+        case TRANSITION_SLIDE_VERTICAL_DOWN: return new VerticalSlideTransition(true);
         case TRANSITION_RANDOM: return transition_type_to_transition( (transition_type_t) ((esp_random() % (TRANSITION_RANDOM - 1)) + 1));
 
         case TRANSITION_NONE:
