+++ src/app/app_host.cpp
@@ -85,7 +85,11 @@ void AppHost::step() {
             push_state(STATE_TIMER_EDITOR, TRANSITION_WIPE);
         }
         else if(hid_test_key_state(KEY_SOFT_F2) == KEYSTATE_HIT) {
-            change_state(STATE_IDLE, TRANSITION_WIPE);
+            if(get_state() == STATE_IDLE) {
+                change_state(STATE_MENU, TRANSITION_SLIDE_HORIZONTAL_RIGHT);
+            } else {
+                change_state(STATE_IDLE, TRANSITION_WIPE);
+            }
         }
         else if(hid_test_key_state(KEY_SOFT_F3) == KEYSTATE_HIT) {
             push_state(STATE_STOPWATCH, TRANSITION_WIPE);
